RENAME #17 WareType
