+++ src/logic/IPlayer.ts
@@ -1,5 +1,5 @@
 import { IWare } from "./IWare";
-import { WareType } from "./wareType";
+import { WareType } from "./WareType";
 
 export interface IPlayer {
     getMoney(): number;
