+++ src/logic/Ware.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import { uniq } from "lodash";
 import { Ware } from "./ware";
-import { WareType } from "./wareType";
+import { WareType } from "./WareType";
 
 describe("Ware", () => {
     it("getWaresOfEachType() has correct length", () => {
