+++ src/logic/logic.integration.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
 import { LogicBuilder } from "./LogicBuilder";
-import { WareType } from "./wareType";
+import { WareType } from "./WareType";
 
 describe("Logic integration", () => {
     it("buy() does not throw", () => {
